refactor - wrap collectorconfig handling with generic informer handlers
